refactor: use path aliases
