+++ client/src/components/TaskTracker/Settings.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
-import { Button } from "../Common/Button";
+import { Button } from "@Components/Common/Button";
 import { AiFillDelete } from "react-icons/ai";
-import { useTask, useMaxPomodoro } from "../../store";
+import { useTask, useMaxPomodoro } from "@Store";
 export const Settings = ({ setOpenSettings, Task }) => {
   const [text, setText] = useState(Task.description);
   const { removeTask, setPomodoro, alertTask } = useTask();
